Refactor remaining actions
